LayoutTests/js/dfg-exception-expected.txt

This tests that exceptions are thrown correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtException is true
PASS successfullyParsed is true

TEST COMPLETE
